LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select/select-keyboard-behavior.optional-expected.txt

one two three
 custom button

Harness Error (FAIL), message = Test named 'defaultbutton: When the listbox is open, the enter key should commit the selected option.' specified 1 'cleanup' function, and 1 failed.

PASS defaultbutton: When the listbox is closed, spacebar should open the listbox.
FAIL defaultbutton: When the listbox is closed, all arrow keys should open the listbox. assert_true: Arrow left should open the listbox. expected true got false
PASS defaultbutton: When the listbox is closed, the enter key should not trigger form submission.
FAIL defaultbutton: When the listbox is open, the enter key should commit the selected option. assert_equals: The selected option should receive initial focus. expected Element node <option class="two">two</option> but got Element node <select id="defaultbutton">
    <option class="one">one</...
NOTRUN defaultbutton: When the listbox is open, the tab key should close the listbox.
NOTRUN custombutton: When the listbox is closed, spacebar should open the listbox.
NOTRUN custombutton: When the listbox is closed, all arrow keys should open the listbox.
NOTRUN custombutton: When the listbox is closed, the enter key should not trigger form submission.
NOTRUN custombutton: When the listbox is open, the enter key should commit the selected option.
NOTRUN custombutton: When the listbox is open, the tab key should close the listbox.
